+++ bigframes/testing/compiler_session.py
