+++ pytest.ini
@@ -2,3 +2,4 @@
 doctest_optionflags = NORMALIZE_WHITESPACE
 filterwarnings =
     ignore::pandas.errors.SettingWithCopyWarning
+addopts = "--import-mode=importlib"
