+++ tests/system/small/engines/test_selection.py
@@ -16,7 +16,7 @@ import pytest
 
 from bigframes.core import array_value, expression, identifiers, nodes
 from bigframes.session import polars_executor
-from tests.system.small.engines.engine_utils import assert_equivalence_execution
+from bigframes.testing.engine_utils import assert_equivalence_execution
 
 pytest.importorskip("polars")
