+++ tests/system/small/regression/test_issue355_merge_after_filter.py
@@ -15,7 +15,7 @@
 import pandas as pd
 import pytest
 
-from tests.system.utils import assert_pandas_df_equal
+from bigframes.testing.utils import assert_pandas_df_equal
 
 
 @pytest.mark.parametrize(
