+++ tests/system/small/test_pandas.py
@@ -21,7 +21,7 @@ import pytest
 import pytz
 
 import bigframes.pandas as bpd
-from tests.system.utils import assert_pandas_df_equal
+from bigframes.testing.utils import assert_pandas_df_equal
 
 
 @pytest.mark.parametrize(
